data_dumper/pyproject.toml
[tool.poetry]
name = "data-dumper"
version = "0.1.0"
description = ""
authors = ["Olivier Neve"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
